Tests/LibWeb/Text/expected/SVG/svg-ownerSVGElement-attribute.txt

svg.ownerSVGElement = 'null'
linearGradient.ownerSVGElement = '[object SVGSVGElement]'
linearGradient.ownerSVGElement == svg = 'true'
